Use `Object.hasOwn` to handle subtests with names like `toString`
This fixes a bug where subtest names like `toString` that are defined on
the prototype (e.g `Function.toString`) makes the logic take the branch
that corresponds to the existence of scores for that subtest. This
causes any subsequent property lookups to be `undefined` which then gets
transformed to NaN and propagates to the final scores.

Signed-off-by: Mukilan Thiyagarajan <mukilan@igalia.com>