Upgrade to latest actions for deploying the site
upload-pages-artifact@v2 uses v3 of upload-artifact which is now deprecated, causing the build to fail with the message

```
This request has been automatically failed because it uses a deprecated version of `actions/upload-artifact: v3`. Learn more: https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/
```

This patch upgrades the workflow to use the latest version of `upload-pages-artifact`. The related `deploy-pages` action also needs to be upgrade as per the release notes:

https://github.com/actions/upload-pages-artifact/releases/tag/v3.0.0

Signed-off-by: Mukilan Thiyagarajan <mukilan@igalia.com>